backend/requirements.txt
alembic==1.7.5
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
config==0.5.1
defusedxml==0.7.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
httpie==2.6.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==8.4.0
Pygments==2.10.0
PyJWT==2.3.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.2
